Work in progress. Save all changes.
git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@2202 b967ad99-d558-0410-b138-e0f6c56caec7

+++ CSPadPixCoords/trunk/include/CSPadImageProducer.h
@@ -146,8 +146,8 @@ private:
   CSPadPixCoords::PixCoords2x1::COORDINATE ZCOOR;
 	
   uint32_t  m_cspad_ind;
-  float    *m_coor_x_pix;
-  float    *m_coor_y_pix;
+  double    *m_coor_x_pix;
+  double    *m_coor_y_pix;
   uint32_t *m_coor_x_int;
   uint32_t *m_coor_y_int;
 
@@ -156,7 +156,7 @@ private:
 
   enum{ NX_CSPAD=1750, 
         NY_CSPAD=1750 };
-  float m_arr_cspad_image[NX_CSPAD][NY_CSPAD];
+  double m_arr_cspad_image[NX_CSPAD][NY_CSPAD];
 };
 
 } // namespace CSPadPixCoords
